requirements.txt
# Application
fastapi==0.103.0
uvicorn==0.22.0

# DB drivers
redis==5.0.3
motor==3.4.0
tornado==6.4

# Validation
pydantic==2.5.3

# Security
PyJWT
passlib
bcrypt

# Utils
python-dotenv